BUG#27583 (slave sql fails to read from iocache when slave got stopped at pos==4):
Submitting patch on behalf of Andrei, who discovered the problem
and provided the patch.

An update of the group relay log coordinates when rotating forgot to
update the group relay log name and only updated the group relay log
position (and group master log name and position).

This patch adds code to update the group relay log *name* as well as
the position