Bug#27044 replicated with unique field ndb table allows duplkey inserts The bug in that slave version of a table with unique field still was able to execute INSERT query as replace whereas it's impossible on master.
The reason of this artifact is wrong usage of ndb->extra:s.

Fixed with resetting flags at do_after.
There is open issue with symmetrical resetting
   table->file->extra(HA_EXTRA_NO_IGNORE_NO_KEY)
which i had to hand to bug#27077.
The test for the current bug was committed in a cset for bug#27320.


sql/log_event.cc:
  fixing do_after_row_operation to reset the effect of the extra engine's 
  flags set at do_before;
  comments on meaning of extra flags added;
  execution of table->file->ha_end_bulk_insert() in do_after is not dependant on error;