BUG#28971 - ALTER TABLE followed by UPDATE for a CSV table             make server crash
UPDATE against CSV table may cause server crash or update a table with wrong
values.

CSV can write only a whole row at once. That means it must read all columns,
that it is not going to update, and write them along with updated columns.
But only limited set of columns was read, those that were needed for the
UPDATE query.

With this fix all columns are read in case we're performing an UPDATE.